Objednávka dialog

+++ App/App/Model/Objednavka.cs
@@ -10,11 +10,11 @@ namespace App.Model
     {
         public int Id { get; set; }
         public DateTime DatumZalozeni { get; set; }
-        public decimal Cena { get; set; }
+        public double Cena { get; set; }
         public int IdZakaznik { get; set; }
         public int? IdFaktura { get; set; }
 
-        public Objednavka(int id, DateTime datumZalozeni, decimal cena, int idZakaznik, int? idFaktura = null)
+        public Objednavka(int id, DateTime datumZalozeni, double cena, int idZakaznik, int? idFaktura = null)
         {
             Id = id;
             DatumZalozeni = datumZalozeni;
